+++ test/Portia.ConsumerTests/ProcessorBaseTests.cs
@@ -44,7 +44,7 @@ public sealed class ProcessorBaseTests
         var services = new ServiceCollection();
         _ = services.AddScoped<Repository>();
         _ = services.AddSingleton<IDomainEventReader, InMemoryEventStore>();
-        _ = services.AddPortia().AddProjector<BatchAccounts>("BatchAccounts", WorkloadScope.PerTenant);
+        _ = services.AddPortia().AddProjector<BatchAccounts>("BatchAccounts", WorkloadScope.Global);
         using var provider = services.BuildServiceProvider(new ServiceProviderOptions
         { ValidateScopes = true, ValidateOnBuild = true });
         using var scope = provider.CreateScope();